requirements-lambda.txt
fastapi==0.122.0
mangum==0.19.0
passlib==1.7.4
python-jose==3.3.0
email-validator==2.3.0
pydantic==2.12.4
pydantic-settings==2.12.0
